Add files via upload
 """
    A class object for computing the value of European call/put options using the binomial tree method and the Black-Scholes formula. 
    The class is instantiated with some predefined variables of S0 = 98, r = 0.06, σ = 0.1, t = 0 and T = 0.5; The class computes the value based on 
    two methods of computing the upward and downward factors once specify in the tree_methods function for the Binomial tree method. 
    "UDP1" uses the Cox-Ross-Rubenstein method whereas the "UDP2" uses the log-normal formula for 
    the upward and downward factors with each associated probability "p".

In this particular notebook, I studied the nature of the error after computing the option value using the Black-Scholes formula and the binomial tree methods. Where in the binomial tree methods two methods were used in defining the upward and downward movements, plotted and observations commented accordingly. 
Feel free to improve on this work to your own satisfaction.
    
    """